+++ packages/elements/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.16](https://github.com/react-navigation/react-navigation/compare/@react-navigation/elements@2.0.0-alpha.15...@react-navigation/elements@2.0.0-alpha.16) (2024-03-22)
+
+**Note:** Version bump only for package @react-navigation/elements
+
+# [2.0.0-alpha.15](https://github.com/react-navigation/react-navigation/compare/@react-navigation/elements@2.0.0-alpha.14...@react-navigation/elements@2.0.0-alpha.15) (2024-03-22)
+
+**Note:** Version bump only for package @react-navigation/elements
+
+# [2.0.0-alpha.14](https://github.com/react-navigation/react-navigation/compare/@react-navigation/elements@2.0.0-alpha.13...@react-navigation/elements@2.0.0-alpha.14) (2024-03-20)
+
+**Note:** Version bump only for package @react-navigation/elements
+
 # [2.0.0-alpha.13](https://github.com/react-navigation/react-navigation/compare/@react-navigation/elements@2.0.0-alpha.12...@react-navigation/elements@2.0.0-alpha.13) (2024-03-14)
 
 ### Bug Fixes